Historiographical_Divergence/requirements.txt
# Core Logic
openai>=1.0.0
anthropic>=0.18.0
google-generativeai>=0.3.0
langchain>=0.1.0

# Scraping & Data
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pandas>=2.0.0
numpy>=1.24.0
pydantic>=2.0.0

# Statistical Analysis
scipy>=1.11.0
scikit-learn>=1.3.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
tqdm>=4.65.0
jupyter>=1.0.0
pytest>=7.4.0
matplotlib>=3.7.0
seaborn>=0.12.0